Frontend Changes
This fixes some little mistakes in LoginJS.js as well as making it so that the Staff creation in 07E-portfolio2.html isn't Staff creation but CNA creation now. There are minor bugs in that second part, such as still having to chose CNA as the position, that will be fixed.